requirements.txt
Flask
requests
tldextract
python-whois
dnspython
cryptography
beautifulsoup4
rapidfuzz
plotly
